Move filter package to feature/ui/filter, infrastructure/web/filter
