+++ airsonic-taglibs/src/main/java/com/tesshu/jpsonic/infrastructure/web/filter/ParameterDecodingFilter.java
@@ -19,7 +19,7 @@
  * (C) 2018 tesshucom
  */
 
-package com.tesshu.jpsonic.filter;
+package com.tesshu.jpsonic.infrastructure.web.filter;
 
 import java.io.IOException;
 import java.util.ArrayList;
