+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/infrastructure/web/filter/RESTFilter.java
@@ -19,7 +19,7 @@
  * (C) 2018 tesshucom
  */
 
-package com.tesshu.jpsonic.filter;
+package com.tesshu.jpsonic.infrastructure.web.filter;
 
 import java.io.IOException;
